Move webview to separate file
