+++ src/ui/ui.ts
@@ -1,5 +1,6 @@
 import * as path from "path";
 import * as vscode from "vscode";
+import {TmcWebview} from "./webview";
 
 /**
  * A class for interacting with the user through graphical means
@@ -40,96 +41,6 @@ export default class UI {
 
         vscode.window.registerTreeDataProvider("tmcView", this.treeDP);
 
-    }
-}
-
-/**
- * A class for handling the Webview component of the plugin UI, to be used through the UI class
- */
-class TmcWebview {
-
-    private extensionContext: vscode.ExtensionContext;
-    private messageHandlers: Map<string, (msg: any) => void> = new Map();
-
-    /**
-     * NOTE: use [[getPanel]] to correctly handle disposed instances
-     */
-    private panel: vscode.WebviewPanel | undefined;
-
-    /**
-     * Creates a TmcWebview object used by the UI class
-     * @param extensionContext The VSCode extension context, required for path resolution for the CSS stylesheet
-     */
-    constructor(extensionContext: vscode.ExtensionContext) {
-        this.extensionContext = extensionContext;
-    }
-
-    /**
-     * Wraps an HTML fragment, representing the body of the document, with a template containing a CSS stylesheet
-     * @param body The HTML fragment, i.e. everything that goes between the body tags
-     */
-    public htmlWrap(body: string): string {
-        return `<html><head><link rel="stylesheet" type="text/css" href="${this.resolvePath("resources/style.css")}"></head><body>${body}</body></html>`;
-    }
-
-    /**
-     * Creates an absolute path to a file in the extension folder, for use within the webview
-     * @param relativePath The relative path to the file within the extension folder, e.g. 'resources/style.css'
-     */
-    public resolvePath(relativePath: string): string {
-        return vscode.Uri.file(path.join(this.extensionContext.extensionPath, relativePath)).toString().replace("file:", "vscode-resource:");
-    }
-
-    /**
-     * Sets the HTML content of the webview and brings it to the front
-     * @param html A string containing a full HTML document, see [[htmlWrap]]
-     */
-    public setContent(html: string) {
-        const panel = this.getPanel();
-        panel.webview.html = html;
-        panel.reveal();
-    }
-
-    /**
-     * Register a handler for a specific message type sent from the webview
-     * @param messageId The message type to handle
-     * @param handler A handler function that receives the full message as a parameter
-     */
-    public registerHandler(messageId: string, handler: (msg: any) => void) {
-        if (this.messageHandlers.get(messageId) !== undefined) {
-            return;
-        }
-        this.messageHandlers.set(messageId, handler);
-    }
-
-    /**
-     * Closes the Webview.
-     */
-    public dispose() {
-        this.panel?.dispose();
-    }
-
-    /**
-     * Re-creates the webview panel if it has been disposed and returns it
-     * @return A webview panel with strong freshness guarantees
-     */
-    private getPanel(): vscode.WebviewPanel {
-        if (this.panel === undefined) {
-            this.panel = vscode.window.createWebviewPanel("tmcmenu", "TestMyCode", vscode.ViewColumn.Active,
-                { enableScripts: true });
-            this.panel.onDidDispose(() => { this.panel = undefined; },
-                this, this.extensionContext.subscriptions);
-            this.panel.webview.onDidReceiveMessage((msg: { type: string, [x: string]: string }) => {
-                const handler = this.messageHandlers.get(msg.type);
-                if (handler) {
-                    handler(msg);
-                } else {
-                    console.error("Unhandled message type: " + msg.type);
-                }
-            },
-                this, this.extensionContext.subscriptions);
-        }
-        return this.panel;
     }
 }
